requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
amqp==1.4.9
anyjson==0.3.3
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via packaging, pytest
billiard==3.3.0.23
celery==3.1.26.post2
click==7.0                # via code-annotations
code-annotations==0.3.2
configparser==3.7.4       # via importlib-metadata
contextlib2==0.5.5        # via importlib-metadata
coverage==4.5.4           # via pytest-cov
ddt==1.2.1
django-celery==3.3.0
django-crum==0.7.3
django-model-utils==3.2.0
djangorestframework==3.9.4
freezegun==0.3.12
funcsigs==1.0.2           # via mock, pytest
importlib-metadata==0.19  # via pluggy, pytest
jinja2==2.10.1            # via code-annotations
kombu==3.0.37
markupsafe==1.1.1         # via jinja2
mock==3.0.5
more-itertools==5.0.0     # via pytest
packaging==19.1           # via pytest
pathlib2==2.3.4           # via importlib-metadata, pytest, pytest-django
pbr==5.4.2                # via stevedore
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==4.6.5             # via pytest-cov, pytest-django
python-dateutil==2.8.0    # via freezegun
python-slugify==3.0.3     # via code-annotations
pytz==2019.2
pyyaml==5.1.2             # via code-annotations
scandir==1.10.0           # via pathlib2
six==1.12.0               # via freezegun, mock, more-itertools, packaging, pathlib2, pytest, python-dateutil, stevedore
stevedore==1.30.1         # via code-annotations
text-unidecode==1.2       # via python-slugify
wcwidth==0.1.7            # via pytest
zipp==0.5.2               # via importlib-metadata
